Minimal Viable Product for @mentions
In the comments this was not working and so left as a TODO. However, the use of a user name (without the `@`) does work. So why not just drop the `@` and proceed as normal. This way a user can enter @mentions if they want to or not. Either way the script works and everyone is happy.